Refactored categories into separate files
